+++ mawabot/cogs/text/core.py
@@ -1,5 +1,5 @@
 #
-# cogs/text.py
+# cogs/text/core.py
 #
 # mawabot - Maware's selfbot
 # Copyright (c) 2017 Ma-wa-re, Ammon Smith
@@ -19,7 +19,7 @@ from discord.ext import commands
 import upsidedown
 
 __all__ = [
-    'setup',
+    'Text',
 ]
 
 class Text:
@@ -37,7 +37,7 @@ class Text:
         fut = ctx.message.delete()
         count = 0
         async for msg in ctx.channel.history():
-            if msg.author != self.bot.client:
+            if msg.author == self.bot.client:
                 count += 1
                 if count >= posts_back + 1:
                     break
@@ -46,7 +46,6 @@ class Text:
         if not msg.content.startswith('.\n'):
             content = '.\n' + msg.content
             await msg.edit(content=content)
-
         await fut
 
     @commands.command()
@@ -64,49 +63,7 @@ class Text:
         await ctx.message.edit(content=result)
 
     @commands.command()
-    async def sw(self, ctx, *, text: str):
-        ''' Spaces out words for meme emphasis '''
-
-        words = text.split(' ')
-        result = []
-        for word in words:
-            result.append(' '.join(word))
-
-        await ctx.message.edit(content=' . '.join(result))
-
-    @commands.command()
-    async def cw(self, ctx, *, text: str):
-        ''' "Crossword"-ifys the given text for meme emphasis '''
-
-        text = text.upper()
-        lines = [text] + list(text[1:])
-
-        await ctx.message.edit(content='\n'.join(lines))
-
-    @commands.command()
     async def rev(self, ctx, *, text: str):
         ''' Reverses the text given '''
 
         await ctx.message.edit(content=text[::-1])
-
-    @commands.command()
-    async def kerrhau(self, ctx, *text: str):
-        ''' "kerrhau"-ifys the given text for meme emphasis '''
-
-        text = list(text)
-        words = []
-
-        while text:
-            word = []
-
-            for _ in range(randint(1, 3)):
-                if text:
-                    word.append(text.pop(0))
-
-            words.append(' '.join(word))
-
-        last = words[-1][-1]
-        words[-1] = words[-1][:-1]
-        words.append(last)
-
-        await ctx.message.edit(content='\n'.join(words))
